Solution to Find the Sum of Digits of Each Element in an Array

+++ Find the Sum of Digits of Each Element in an Array@@676c02002dbc8c3e7a280b7f/code.c
@@ -13,7 +13,7 @@ int main(){
         while(b!=0){
             rem = b%10;
             count += rem;
-            b = int(b/10);
+            b = (b/10);
         }
         sod[i]=count;
     }
